backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
aiosqlite==0.22.1
asyncpg==0.31.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.26
pydantic-settings==2.13.1
httpx==0.28.1
supabase==2.28.3
anthropic==0.95.0
PyMuPDF==1.27.2.2
redis==7.4.0
python-dotenv==1.2.2
